+++ src/features/topic/topic.controller.ts
@@ -1,9 +1,10 @@
 import { Response, Request } from "express";
 import * as topicService from './topic.service'
-import { IUpdateData } from "../helpers/update.interface";
+import { IUpdateData } from "../utils/update.interface";
 import Topic from "./topic.model";
 import ITopic from "./topic.interface";
 import { Types } from "mongoose";
+import { handelValidation } from "../utils/common.validators";
 async function getTopicsController(req: Request, res: Response): Promise<void> {
     try{
         const topics = await topicService.getTopicsService()
@@ -15,8 +16,14 @@ async function getTopicsController(req: Request, res: Response): Promise<void> {
 
 async function getTopicController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const id = req.params.topic_id
         const topic = await topicService.getTopicService(new Types.ObjectId(id));
+        if (!topic) throw new Error('Topic not found');
         res.json(topic)
     }catch(err: any){
         res.status(404).json(err.message);
@@ -25,6 +32,11 @@ async function getTopicController(req: Request, res: Response): Promise<void> {
 
 async function createTopicController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const topicData: ITopic = {
             title: reqBody.title,
@@ -43,6 +55,11 @@ async function createTopicController(req: Request, res: Response): Promise<void>
 
 async function updateTopicController(req: Request, res: Response): Promise<void> {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const reqBody = req.body;
         const {topic_id} = req.params;
         const data: IUpdateData = {
@@ -58,8 +75,13 @@ async function updateTopicController(req: Request, res: Response): Promise<void>
 }
 
 async function deleteTopicController(req: Request, res: Response): Promise<void> {
-    const topicId = new Types.ObjectId(req.params.topic);
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
+        const topicId = new Types.ObjectId(req.params.topic);
         const val = await topicService.deleteTopicService(topicId);
         if (val)
             res.sendStatus(204);
@@ -72,6 +94,11 @@ async function deleteTopicController(req: Request, res: Response): Promise<void>
 
 function addFavoredUserController(req: Request, res: Response): void {
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const {topic_id} = req.params;
         const {userId} = req.body;
         topicService.addFavoredUserService(new Types.ObjectId(topic_id), new Types.ObjectId(userId));
@@ -82,6 +109,11 @@ function addFavoredUserController(req: Request, res: Response): void {
 }
 function removeFavoredUserController(req: Request, res: Response): void{
     try{
+        const errors = handelValidation(req);
+        if (errors.length > 0){
+            res.status(400).json(errors);
+            return;
+        }
         const {topic_id, user_id} = req.params;
         topicService.removeFavoredUserService(new Types.ObjectId(topic_id), new Types.ObjectId(user_id));
         res.sendStatus(204);
